requirements.txt
# LLM / HTTP
requests>=2.31.0
python-dotenv>=1.0.0

# データ処理
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# テクニカル分析
# ta-lib>=0.4.28  # TA-Lib はバイナリが必要。Dockerfileで apt-get インストール後に有効化

# データ取得・スクレイピング
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0

# 型・バリデーション
pydantic>=2.5.0
yfinance>=0.2.40
openpyxl>=3.1.0
xlrd>=2.0.1
